+++ docs/spie2022_radial_subsampling.rst
@@ -1,0 +1,1 @@
+.. include:: ../projects/spie2022_radial_subsampling/README.rst
