### 2022-03-11 e01c644
Delete projects/spie2022_radial_subsampling directory
