+++ projects/spie2022_radial_subsampling/README.rst
@@ -51,7 +51,7 @@ To make predictions on the validation set (14 volumes, see `lists/val/ <https://
 
 .. code-block:: bash
 
-    cd projects/spie2022_radial_subsampling/
+    cd projects/
     python3 predict_val.py <data_root>/Val/ \
                     <output_directory> \
                     <experiment_directory_containing_checkpoint> \
