+++ projects/cvpr2022_recurrentvarnet/configs/comparisons/base_rim.yaml
@@ -1,135 +1,0 @@
-physics:
-    forward_operator: fft2(centered=False)
-    backward_operator: ifft2(centered=False)
-training:
-    datasets:
-        # Two datasets, only difference is the shape, so the data can be collated for larger batches
-        -   name: CalgaryCampinas
-            lists:
-                - ../../lists/train/12x218x170_train.lst
-            transforms:
-                crop: null
-                estimate_sensitivity_maps: true  # Estimate the sensitivity map on the ACS
-                scaling_key: masked_kspace  # Compute the image normalization based on the masked_kspace maximum
-                image_center_crop: false
-                masking:
-                    name: CalgaryCampinas
-                    accelerations: [5, 10]
-            crop_outer_slices: true
-        -   name: CalgaryCampinas
-            lists:
-                - ../../lists/train/12x218x180_train.lst
-            transforms:
-                crop: null
-                estimate_sensitivity_maps: true  # Estimate the sensitivity map on the ACS
-                scaling_key: masked_kspace  # Compute the image normalization based on the masked_kspace maximum
-                image_center_crop: false
-                masking:
-                    name: CalgaryCampinas
-                    accelerations: [5, 10]
-            crop_outer_slices: true
-    batch_size: 1  # This is the batch size per GPU!
-    optimizer: Adam
-    lr: 0.0005
-    weight_decay: 0.0
-    lr_step_size: 20000
-    lr_gamma: 0.2
-    lr_warmup_iter: 1000
-    num_iterations: 500000
-    gradient_steps: 1
-    gradient_clipping: 0.0
-    gradient_debug: false
-    checkpointer:
-        checkpoint_steps: 500
-    validation_steps: 1000 # With batch size 4 and 4 GPUs this is about 7300 iterations, or ~1 epoch.
-    loss:
-        crop: null
-        losses:
-            -   function: l1_loss
-                multiplier: 1.0
-            -   function: ssim_loss
-                multiplier: 1.0
-validation:
-    datasets:
-        # Twice the same dataset but a different acceleration factor
-        -   name: CalgaryCampinas
-            lists:
-                - ../../lists/val/12x218x170_val.lst
-                - ../../lists/val/12x218x180_val.lst
-            transforms:
-                crop: null
-                estimate_sensitivity_maps: true
-                scaling_key: masked_kspace
-                masking:
-                    name: CalgaryCampinas
-                    accelerations: [5]
-            crop_outer_slices: true
-            text_description: 5x  # Description for logging
-        -   name: CalgaryCampinas
-            lists:
-                - ../../lists/val/12x218x170_val.lst
-                - ../../lists/val/12x218x180_val.lst
-            transforms:
-                crop: null
-                estimate_sensitivity_maps: true
-                scaling_key: masked_kspace
-                masking:
-                    name: CalgaryCampinas
-                    accelerations: [10]
-            crop_outer_slices: true
-            text_description: 10x  # Description for logging
-    crop: null  # This sets the cropping for the DoIterationOutput
-    metrics:  # These are obtained from direct.functionals
-        - calgary_campinas_psnr
-        - calgary_campinas_ssim
-        - fastmri_nmse
-model:
-    model_name: rim.rim.RIM
-    hidden_channels: 128
-    image_initialization: sense  # This uses the computed sensitivity map to create a zero-filled reconstruction
-    length: 16
-    depth: 2
-    steps: 1
-    no_parameter_sharing: false
-    instance_norm: false
-    dense_connect: false
-    replication_padding: true
-additional_models:
-    sensitivity_model:
-        model_name: unet.unet_2d.UnetModel2d
-        in_channels: 2
-        out_channels: 2
-        num_filters: 8
-        num_pool_layers: 4
[…]
