+++ projects/cvpr2022_recurrentvarnet/README.rst
@@ -5,8 +5,8 @@ Recurrent Variational Network: A Deep Learning Inverse Problem Solver applied to
 This folder contains the training code specific for our experiments presented in our paper
 `Recurrent Variational Network: A Deep Learning Inverse Problem Solver applied to the task of Accelerated MRI Reconstruction <https://arxiv.org/abs/2111.09639>`__ (pre-print version) accepted in CVPR 2022.
 
-Dataset
--------
+Datasets
+--------
 * For the proposed model, the comparison, and ablation studies we used the `Calgary-Campinas public brain multi-coil MRI dataset <https://sites.google.com/view/calgary-campinas-dataset/home>`__ which was released as part of an accelerated MRI reconstruction challenge. The dataset is consisted of 67  3D raw k-space volumes. After cropping the 100 outer slices, these amount to 10,452 slices of fully sampled k-spaces which we randomly split into training (47 volumes), validation (10 volumes) and test (10 volumes) sets (see `lists/ <https://github.com/NKIAI/direct/tree/main/projects/cvpr2022_recurrentvarnet/calgary_campinas/lists>`__). Sub-sampling was performed by applying the Poisson disk distribution sub-sampling masks provided by the challange.
 
 * For additional experiments we used the AXT1 brain `FastMRI dataset <https://fastmri.org/dataset/>`_. The dataset was consisted of 3D raw k-space volumes:
