+++ paper.bib
@@ -57,7 +57,7 @@ volume = {53},
 pages = {64-78},
 year = {2019},
 issn = {1361-8415},
-doi = {https://doi.org/10.1016/j.media.2019.01.005},
+doi = {10.1016/j.media.2019.01.005},
 url = {https://www.sciencedirect.com/science/article/pii/S1361841518306078},
 author = {Kai Lønning and Patrick Putzky and Jan-Jakob Sonke and Liesbeth Reneman and Matthan W.A. Caan and Max Welling},
 keywords = {MRI, Reconstruction, Deep learning, Inverse problems},
@@ -96,7 +96,7 @@ volume = {80},
 number = {5},
 pages = {2188-2201},
 keywords = {convolutional neural networks, cross-domain deep learning, image reconstruction, k-space completion, MRI acceleration},
-doi = {https://doi.org/10.1002/mrm.27201},
+doi = {10.1002/mrm.27201},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/mrm.27201},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/mrm.27201},
 year = {2018}
@@ -106,6 +106,7 @@ year = {2018}
     author    = {Jun, Yohan and Shin, Hyungseob and Eo, Taejoon and Hwang, Dosik},
     title     = {Joint Deep Model-Based MR Image and Coil Sensitivity Reconstruction Network (Joint-ICNet) for Fast MRI},
     booktitle = {Proceedings of the IEEE/CVF Conference on Computer Vision and Pattern Recognition (CVPR)},
+    doi       = {10.1109/cvpr46437.2021.00523},
     month     = {June},
     year      = {2021},
     pages     = {5270-5279}
