### 2022-03-28 #4278 @editorialbot
[REVIEW]: DIRECT: Deep Image REConstruction Toolkit
**Submitting author:** @georgeyiasemis (<a href="http://orcid.org/0000-0002-1348-8987">George Yiasemis</a>)
**Repository:** https://github.com/NKI-AI/direct
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v1.0.1
