### 2022-03-29 #4239 @zaccharieramzi
> @zaccharieramzi would you be interested in reviewing this submission to the Journal of Open Source Software?

Hi @osorensen ,

I am currently on holiday with little internet access. I think the submission could be in my range of expertise, in which case I would be interested in reviewing it. But can I get back to you after April 3rd to give you a final answer?
