### 2022-03-29 7c9d4f5
Update checkpointer.py
New version of `pylint` complaints (modified-iterating-dict)
