### 2022-03-29 933f7af
Update checkpointer.py
Codacy complaints.
