### 2022-03-31 966c550
header crop for inference
