+++ projects/cvpr2022_recurrentvarnet/fastmri/AXT1_brain/configs_inference/4x/base_lpd.yaml
@@ -117,3 +117,4 @@ inference:
                 accelerations: [4]
                 center_fractions: [0.08]
         text_description: inference-4x  # Description for logging
+    crop: header
