### 2022-03-31 8ba322a
Add inference configs
