+++ projects/cvpr2022_recurrentvarnet/calgary_campinas/configs_inference/5x/comparisons/base_unet.yaml
@@ -6,7 +6,7 @@ training:
         # Two datasets, only difference is the shape, so the data can be collated for larger batches
         -   name: CalgaryCampinas
             lists:
-                - ../lists/train/12x218x170_train.lst
+                - ../../lists/train/12x218x170_train.lst
             transforms:
                 crop: null
                 estimate_sensitivity_maps: true  # Estimate the sensitivity map on the ACS
@@ -18,7 +18,7 @@ training:
             crop_outer_slices: true
         -   name: CalgaryCampinas
             lists:
-                - ../lists/train/12x218x180_train.lst
+                - ../../lists/train/12x218x180_train.lst
             transforms:
                 crop: null
                 estimate_sensitivity_maps: true  # Estimate the sensitivity map on the ACS
@@ -28,11 +28,11 @@ training:
                     name: CalgaryCampinas
                     accelerations: [5, 10]
             crop_outer_slices: true
-    batch_size: 4  # This is the batch size per GPU!
+    batch_size: 32  # This is the batch size per GPU!
     optimizer: Adam
-    lr: 0.0005
+    lr: 0.0002
     weight_decay: 0.0
-    lr_step_size: 20000
+    lr_step_size: 10000
     lr_gamma: 0.2
     lr_warmup_iter: 1000
     num_iterations: 500000
@@ -40,7 +40,7 @@ training:
     gradient_clipping: 0.0
     gradient_debug: false
     checkpointer:
-        checkpoint_steps: 1000
+        checkpoint_steps: 500
     validation_steps: 1000
     loss:
         crop: null
@@ -54,8 +54,8 @@ validation:
         # Twice the same dataset but a different acceleration factor
         -   name: CalgaryCampinas
             lists:
-                - ../lists/val/12x218x170_val.lst
-                - ../lists/val/12x218x180_val.lst
+                - ../../lists/val/12x218x170_val.lst
+                - ../../lists/val/12x218x180_val.lst
             transforms:
                 crop: null
                 estimate_sensitivity_maps: true
@@ -67,8 +67,8 @@ validation:
             text_description: 5x  # Description for logging
         -   name: CalgaryCampinas
             lists:
-                - ../lists/val/12x218x170_val.lst
-                - ../lists/val/12x218x180_val.lst
+                - ../../lists/val/12x218x170_val.lst
+                - ../../lists/val/12x218x180_val.lst
             transforms:
                 crop: null
                 estimate_sensitivity_maps: true
@@ -84,15 +84,9 @@ validation:
         - calgary_campinas_ssim
         - fastmri_nmse
 model:
-    model_name: recurrentvarnet.recurrentvarnet.RecurrentVarNet
-    num_steps: 8
-    recurrent_hidden_channels: 128
-    recurrent_num_layers: 4
-    initializer_initialization: sense
-    learned_initializer: true
-    initializer_channels: [32, 32, 64, 64]
-    initializer_dilations: [1, 1, 2, 4]
-    initializer_multiscale: 3
+    model_name: unet.unet_2d.Unet2d
+    num_filters: 64
+    image_initialization: sense
 additional_models:
     sensitivity_model:
         model_name: unet.unet_2d.UnetModel2d
@@ -105,30 +99,15 @@ logging:
     tensorboard:
         num_images: 4
 inference:
-    datasets:
-        # Twice the same dataset but a different acceleration factor
-        - name: CalgaryCampinas
-          lists:
-              - ../lists/test/12x218x170_test.lst
-          transforms:
-              crop: null
-              estimate_sensitivity_maps: true
-              scaling_key: masked_kspace
-              masking:
-                  name: CalgaryCampinas
-                  accelerations: [ 5 ]
-          crop_outer_slices: true
-          text_description: inference_5x  # Description for logging
[…]
