+++ projects/cvpr2022_recurrentvarnet/calgary_campinas/configs/ablation/base_recurrentvarnet_noRSI.yaml
@@ -100,31 +100,3 @@ additional_models:
 logging:
     tensorboard:
         num_images: 4
-inference:
-    datasets:
-        # Twice the same dataset but a different acceleration factor
-        - name: CalgaryCampinas
-          lists:
-              - ../../lists/test/12x218x170_test.lst
-          transforms:
-              crop: null
-              estimate_sensitivity_maps: true
-              scaling_key: masked_kspace
-              masking:
-                  name: CalgaryCampinas
-                  accelerations: [ 5 ]
-          crop_outer_slices: true
-          text_description: inference_5x  # Description for logging
-        - name: CalgaryCampinas
-          lists:
-              - ../../lists/test/12x218x170_test.lst
-          transforms:
-              crop: null
-              estimate_sensitivity_maps: true
-              scaling_key: masked_kspace
-              masking:
-                  name: CalgaryCampinas
-                  accelerations: [ 10 ]
-          crop_outer_slices: true
-          text_description: inference_10x  # Description for logging
-    batch_size: 8

[…]
