### 2022-04-02 #4278 @sairamgeethanath
@georgeyiasemis - Could you please let me know if there is a way for me to access the Docker image?
