### 2022-04-02 #4278 @georgeyiasemis
>
Hi @sairamgeethanath this is it:
https://github.com/NKI-AI/direct/tree/main/docker
