### 2022-04-02 3be1ffe
Add docker info in installation
