### 2022-04-02 23593f4
Update checkpointer.py
New `pylint` version complaints
