### 2022-04-04 a54077e
Update documentation (#183)
Minor documentation updates:
* Fix broken link in installation.rst
* docker/README.md -> /README.rst
