### 2022-04-04 6deb1cd
Update documentation (#183)
Minor documentation updates:
* Fix broken link in installation.rst
* docker/README.md -> /README.rst
