+++ direct/functionals/ssim.py
@@ -34,8 +34,8 @@ class SSIMLoss(nn.Module):
         super().__init__()
         self.win_size = win_size
         self.k1, self.k2 = k1, k2
-        self.register_buffer("w", torch.ones(1, 1, win_size, win_size) / win_size**2)
-        NP = win_size**2
+        self.register_buffer("w", torch.ones(1, 1, win_size, win_size) / win_size ** 2)
+        NP = win_size ** 2
         self.cov_norm = NP / (NP - 1)
 
     def forward(self, X, Y, data_range):
@@ -53,7 +53,7 @@ class SSIMLoss(nn.Module):
         A1, A2, B1, B2 = (
             2 * ux * uy + C1,
             2 * vxy + C2,
-            ux**2 + uy**2 + C1,
+            ux ** 2 + uy ** 2 + C1,
             vx + vy + C2,
         )
         D = B1 * B2
