### 2022-04-04 4fea3e7
`modulus` takes as input complex_axis and use `modulus` instead of `modulus_if_complex`
