+++ direct/engine.py
@@ -301,6 +301,7 @@ class Engine(ABC, DataDimensionality):
             try:
                 iteration_output = self._do_iteration(data, loss_fns, regularizer_fns=regularizer_fns)
                 output = iteration_output.output_image
+                print("ENGINE", output.shape)
                 loss_dict = iteration_output.data_dict
             except (ProcessKilledException, TrainingException) as e:
                 # If the process is killed, the DoIterationOutput
@@ -370,7 +371,7 @@ class Engine(ABC, DataDimensionality):
 
             metrics_dict = evaluate_dict(
                 metric_fns,
-                T.modulus_if_complex(output.detach()),
+                output.detach(),
                 data["target"].detach().to(self.device),
                 reduction="mean",
             )
