### 2022-04-04 3660b3a
Remove print statement
