### 2022-04-04 7d8f023
Remove `modulus_if_complex`
