+++ paper.bib
@@ -1,7 +1,7 @@
 @misc{yiasemis2021deep,
       title={Deep MRI Reconstruction with Radial Subsampling},
       author={George Yiasemis and Chaoping Zhang and Clara I. Sánchez and Jan-Jakob Sonke and Jonas Teuwen},
-      year={2021},
+      year={2021a},
       eprint={2108.07619},
       archivePrefix={arXiv},
       primaryClass={eess.IV}
@@ -124,7 +124,7 @@ year = {2018}
 @misc{yiasemis2021recurrent,
       title={Recurrent Variational Network: A Deep Learning Inverse Problem Solver applied to the task of Accelerated MRI Reconstruction}, 
       author={George Yiasemis and Clara I. Sánchez and Jan-Jakob Sonke and Jonas Teuwen},
-      year={2021},
+      year={2021b},
       eprint={2111.09639},
       archivePrefix={arXiv},
       primaryClass={eess.IV}
