+++ paper.bib
@@ -174,3 +174,26 @@ year = {2018}
     publisher = {Curran Associates, Inc.},
     url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf}
 }
+
+@article{Lustig2007,
+  doi = {10.1002/mrm.21391},
+  url = {https://doi.org/10.1002/mrm.21391},
+  year = {2007},
+  publisher = {Wiley},
+  volume = {58},
+  number = {6},
+  pages = {1182--1195},
+  author = {Michael Lustig and David Donoho and John M. Pauly},
+  title = {Sparse {MRI}: The application of compressed sensing for rapid {MR} imaging},
+  journal = {Magnetic Resonance in Medicine}
+}
+
+@ARTICLE{1580791,
+  author={Candes, E.J. and Romberg, J. and Tao, T.},
+  journal={IEEE Transactions on Information Theory}, 
+  title={Robust uncertainty principles: exact signal reconstruction from highly incomplete frequency information}, 
+  year={2006},
+  volume={52},
+  number={2},
+  pages={489-509},
+  doi={10.1109/TIT.2005.862083}}
