+++ paper.bib
@@ -197,3 +197,31 @@ year = {2018}
   number={2},
   pages={489-509},
   doi={10.1109/TIT.2005.862083}}
+
+@article{Griswold2002,
+  doi = {10.1002/mrm.10171},
+  url = {https://doi.org/10.1002/mrm.10171},
+  year = {2002},
+  month = jun,
+  publisher = {Wiley},
+  volume = {47},
+  number = {6},
+  pages = {1202--1210},
+  author = {Mark A. Griswold and Peter M. Jakob and Robin M. Heidemann and Mathias Nittka and Vladimir Jellus and Jianmin Wang and Berthold Kiefer and Axel Haase},
+  title = {Generalized autocalibrating partially parallel acquisitions ({GRAPPA})},
+  journal = {Magnetic Resonance in Medicine}
+}
+
+@article{Pruessmann1999,
+  doi = {10.1002/(sici)1522-2594(199911)42:5<952::aid-mrm16>3.0.co;2-s},
+  url = {https://doi.org/10.1002/(sici)1522-2594(199911)42:5<952::aid-mrm16>3.0.co;2-s},
+  year = {1999},
+  month = nov,
+  publisher = {Wiley},
+  volume = {42},
+  number = {5},
+  pages = {952--962},
+  author = {Klaas P. Pruessmann and Markus Weiger and Markus B. Scheidegger and Peter Boesiger},
+  title = {{SENSE}: Sensitivity encoding for fast {MRI}},
+  journal = {Magnetic Resonance in Medicine}
+}
