+++ paper.md
@@ -59,15 +59,15 @@ A plethora of image processing problems arising in biology, chemistry and medici
 where $\mathcal{A}$ denotes the forward operator and $\vec{n}$ is some measurement noise, often assumed to be additive and normally distributed. Equation \ref{eq:eq1} is usually ill-posed and therefore an explicit solution is hard to find. Instead, inverse problems in imaging are typically solved by minimizing an objective function $\mathcal{J}$ which is consisted of a data-fidelity term $\mathcal{L}$ and a regularization term $\mathcal{R}$ (also known as Variational Problems):
     
 \begin{equation}
-    \vec{\hat{x}} \, = \, \min_{\vec{z} \, \in \, \mathcal{X}} \mathcal{J}(z) \, = \, \min_{\vec{z} \, \in \,  \mathcal{X}} \mathcal{L}\big( \, \vec{y}, \, \mathcal{A}(\vec{z})\big) \,+\, \lambda \mathcal{R}(\vec{z}),\quad \lambda \, \ge \, 0.
+    \vec{\hat{x}} \, = \, \min_{\vec{z} \, \in \, \mathcal{X}} \mathcal{J}(z) \, = \, \min_{\vec{z} \, \in \,  \mathcal{X}} \big|\big| \, \vec{y}, \, \mathcal{A}(\vec{z})\big|\big|_2^2 \,+\, \lambda \mathcal{R}(\vec{z}),\quad \lambda \, \ge \, 0.
     \label{eq:eq2}
 \end{equation}
 
-## Accelerated MRI Reconstruction
+## Accelerated Parallel MRI Reconstruction
 
-Accelerated Magnetic Resonance Image (MRI) Reconstruction, that is, reconstructing an MR image from a set of partially observed (or sub-sampled) $k$-space measurements from multiple receiver coils, is par excellence an example of inverse problems. The base forward operator of Accelerated MRI Reconstruction is usually the two or three-dimensional Fast Fourier Transform (FFT) denoted as $\mathcal{F}$.  Conventional approaches for solving this class of inverse problems include Parallel Imaging (PI) [@Larkman_2007] and Compressed Sensing (CS) [@1614066]. Combining these methods with Deep Learning imaging inverse problem solvers can aid in providing reconstructed images with high fidelity from highly sub-sampled measurements. 
+Accelerated Parallel Magnetic Resonance Image (MRI) Reconstruction, that is, reconstructing an MR image from a set of partially observed (or sub-sampled) $k$-space measurements from multiple receiver coils (Parallel Imaging [@Larkman_2007]), is par excellence an example of inverse problems. The base forward operator of Accelerated MRI Reconstruction is usually the two or three-dimensional Fast Fourier Transform (FFT) denoted as $\mathcal{F}$.  Conventional approaches for solving this class of inverse problems include Compressed Sensing (CS) [@1614066, @1580791, @Lustig2007], SENSE [@Pruessmann1999], and GRAPPA [@Griswold200]. Deep Learning-based imaging inverse problem solvers have shown to outperform these conventional techniques by outputting reconstructed images with higher fidelity from highly sub-sampled measurements [@]. 
      
-More specifically, given as input (retrospectively) sub-sampled $k$-space measurements from $n_c$ coils
+More specifically, given as input (retrospectively) sub-sampled $k$-space measurements from $n_c$ receiver coils
 
 \begin{equation*}
[…]
