+++ paper.bib
@@ -225,3 +225,28 @@ year = {2018}
   title = {{SENSE}: Sensitivity encoding for fast {MRI}},
   journal = {Magnetic Resonance in Medicine}
 }
+
+@article{Knoll2020,
+  doi = {10.1109/msp.2019.2950640},
+  url = {https://doi.org/10.1109/msp.2019.2950640},
+  year = {2020},
+  month = jan,
+  publisher = {Institute of Electrical and Electronics Engineers ({IEEE})},
+  volume = {37},
+  number = {1},
+  pages = {128--140},
+  author = {Florian Knoll and Kerstin Hammernik and Chi Zhang and Steen Moeller and Thomas Pock and Daniel K. Sodickson and Mehmet Akcakaya},
+  title = {Deep-Learning Methods for Parallel Magnetic Resonance Imaging Reconstruction: A Survey of the Current Approaches,  Trends,  and Issues},
+  journal = {{IEEE} Signal Processing Magazine}
+}
+
+@article{arxiv.2109.08618,
+  doi = {10.48550/ARXIV.2109.08618},
+  url = {https://arxiv.org/abs/2109.08618},
+  author = {Pal, Arghya and Rathi, Yogesh},
+  keywords = {Image and Video Processing (eess.IV), Computer Vision and Pattern Recognition (cs.CV), Machine Learning (cs.LG), FOS: Electrical engineering, electronic engineering, information engineering, FOS: Electrical engineering, electronic engineering, information engineering, FOS: Computer and information sciences, FOS: Computer and information sciences},
+  title = {A review and experimental evaluation of deep learning methods for MRI reconstruction},  
+  publisher = {arXiv},
+  year = {2021},
+  copyright = {arXiv.org perpetual, non-exclusive license}
+}
