### 2022-04-05 NKI-AI/direct#191 @estenhl
Problems building docker image
Unable to build the docker image on both Ubuntu and macOS due to what seems like multiple errors

**To Reproduce**
````
> docker build -t direct:latest .
Sending build context to Docker daemon  6.144kB
Error response from daemon: dockerfile parse error line 21: unknown instruction: &&
````
Fixed by adding backslashes to lines 20-24
````
 => ERROR [10/15] RUN jupyter notebook --generate-config                   0.3s
------
 > [10/15] RUN jupyter notebook --generate-config:
#14 0.259 /bin/sh: 1: jupyter: not found
````
Fixed by adding a line for conda installing jupyter
````
 => ERROR [14/15] RUN python -m pip install -e ".[dev]"                    1.2s
------
 > [14/15] RUN python -m pip install -e ".[dev]":
#18 0.890 ERROR: File "setup.py" or "setup.cfg" not found. Directory cannot be installed in editable mode: /direct
````
Wasn't able to get to the bottom of this one

**Environment**
Operating system: Ubuntu 18.04.5 LTS
Docker version: 20.10.14

and

Operating system: macOS Monterey 12.2.1
Docker version: 20.10.8
