### 2022-04-05 NKI-AI/direct#191 @georgeyiasemis
Hi @estenhl, can you follow the instructions using `conda` (https://docs.aiforoncology.nl/direct/installation.html#install-using-conda), as we try to solve the above issue?
