### 2022-04-05 NKI-AI/direct#191 @sairamgeethanath
@georgeyiasemis I think lines 20 through 24 are missing new line characters (\)

RUN git clone https://github.com/mrirecon/bart.git /tmp/bart
    && cd /tmp/bart
    && make -j4
    && make install
    && rm -rf /tmp/bart
