+++ projects/cvpr2022_recurrentvarnet/fastmri/AXT1_brain/configs/base_recurrentvarnet_ablation.yaml
@@ -1,0 +1,114 @@
+physics:
+    forward_operator: fft2
+    backward_operator: ifft2
+training:
+    datasets:
+        -   name: FastMRI
+            lists:
+                - ../lists/train.lst
+            transforms:
+                crop: null
+                estimate_sensitivity_maps: true  # Estimate the sensitivity map on the ACS
+                scaling_key: masked_kspace  # Compute the image normalization based on the masked_kspace maximum
+                image_center_crop: false
+                masking:
+                    name: FastMRIRandom
+                    accelerations: [4]
+                    center_fractions: [0.08]
+            crop_outer_slices: false
+        -   name: FastMRI
+            lists:
+                - ../lists/train.lst
+            transforms:
+                crop: null
+                estimate_sensitivity_maps: true  # Estimate the sensitivity map on the ACS
+                scaling_key: masked_kspace  # Compute the image normalization based on the masked_kspace maximum
+                image_center_crop: false
+                masking:
+                    name: FastMRIRandom
+                    accelerations: [8]
+                    center_fractions: [0.04]
+            crop_outer_slices: false
+    batch_size: 1  # This is the batch size per GPU!
+    optimizer: Adam
+    lr: 0.0005
+    weight_decay: 0.0
+    lr_step_size: 30000
+    lr_gamma: 0.2
+    lr_warmup_iter: 100
+    num_iterations: 500000
+    gradient_steps: 1
+    gradient_clipping: 0.0
+    gradient_debug: false
+    checkpointer:
+        checkpoint_steps: 500
+    validation_steps: 2000
+    loss:
+        crop: header
+        losses:
+            -   function: l1_loss
+                multiplier: 1.0
+            -   function: ssim_loss
+                multiplier: 1.0
+validation:
+    datasets:
+        # Twice the same dataset but a different acceleration factor
+        -   name: FastMRI
+            lists:
+                - ../lists/val.lst
+            transforms:
+                crop: null
+                estimate_sensitivity_maps: true
+                scaling_key: masked_kspace
+                masking:
+                    name: FastMRIRandom
+                    accelerations: [4]
+                    center_fractions: [0.08]
+            crop_outer_slices: false
+            text_description: 4x  # Description for logging
+        -   name: FastMRI
+            lists:
+                - ../lists/val.lst
+            transforms:
+                crop: null
+                estimate_sensitivity_maps: true
+                scaling_key: masked_kspace
+                masking:
+                    name: FastMRIRandom
+                    accelerations: [8]
+                    center_fractions: [0.04]
+            text_description: 8x  # Description for logging
+    crop: header  # This sets the cropping for the DoIterationOutput
+    metrics:  # These are obtained from direct.functionals
+        - fastmri_psnr
+        - fastmri_ssim
+model:
+    model_name: recurrentvarnet.recurrentvarnet.RecurrentVarNet
+    num_steps: 6
+    recurrent_hidden_channels: 128
+    recurrent_num_layers: 2
+    initializer_initialization: sense
+    learned_initializer: true
+    initializer_channels: [32, 32, 64, 64]
+    initializer_dilations: [1, 1, 2, 4]
+    initializer_multiscale: 3
+additional_models:
+    sensitivity_model:
+        model_name: unet.unet_2d.UnetModel2d
+        in_channels: 2
+        out_channels: 2
+        num_filters: 8
+        num_pool_layers: 4
+        dropout_probability: 0.0
+logging:
+    tensorboard:
+        num_images: 4
+inference:
+    batch_size: 8
+    dataset:
+        name: FastMRI
+        text_description: inference
+        transforms:
+            crop: null
+            estimate_sensitivity_maps: true
+            scaling_key: masked_kspace

[…]
