+++ direct/nn/mri_models.py
@@ -141,9 +141,7 @@ class MRIModelEngine(Engine):
                 L1 loss.
             """
             resolution = get_resolution(**data)
-            l1_loss = F.l1_loss(
-                *_crop_volume(T.modulus_if_complex(source), data["target"], resolution), reduction=reduction
-            )
+            l1_loss = F.l1_loss(*_crop_volume(source, data["target"], resolution), reduction=reduction)
 
             return l1_loss
 
@@ -165,9 +163,7 @@ class MRIModelEngine(Engine):
                 L2 loss.
             """
             resolution = get_resolution(**data)
-            l2_loss = F.mse_loss(
-                *_crop_volume(T.modulus_if_complex(source), data["target"], resolution), reduction=reduction
-            )
+            l2_loss = F.mse_loss(*_crop_volume(source, data["target"], resolution), reduction=reduction)
 
             return l2_loss
 
@@ -194,7 +190,7 @@ class MRIModelEngine(Engine):
                     f"SSIM loss can only be computed with reduction == 'mean'." f" Got reduction == {reduction}."
                 )
 
-            source_abs, target_abs = _crop_volume(T.modulus_if_complex(source), data["target"], resolution)
+            source_abs, target_abs = _crop_volume(source, data["target"], resolution)
             data_range = torch.tensor([target_abs.max()], device=target_abs.device)
 
             ssim_loss = SSIMLoss().to(source_abs.device).forward(source_abs, target_abs, data_range=data_range)
@@ -490,8 +486,6 @@ def _process_output(
     # data is of shape (batch, complex=2, height, width)
     if scaling_factors is not None:
         data = data * scaling_factors.view(-1, *((1,) * (len(data.shape) - 1))).to(data.device)
-
-    data = T.modulus_if_complex(data)
 
     if len(data.shape) == 3:  # (batch, height, width)
         data = data.unsqueeze(1)  # Added channel dimension.
