### 2022-04-08 b5d85df
Improve typing and documentation
