+++ direct/data/mri_transforms.py
@@ -5,7 +5,7 @@
 import functools
 import logging
 import warnings
-from typing import Any, Callable, Dict, Iterable, Optional
+from typing import Any, Callable, Dict, Iterable, List, Optional, Tuple, Union
 
 import numpy as np
 import torch
@@ -24,17 +24,37 @@ class Compose(DirectModule):
     additional dependencies.
     """
 
-    def __init__(self, transforms: Iterable) -> None:
+    def __init__(self, transforms: Iterable[Callable]) -> None:
+        """Inits :class:`Compose`.
+
+        Parameters
+        ----------
+        transforms: Iterable[Callable]
+            List of transforms.
+        """
         super().__init__()
         self.transforms = transforms
 
-    def __call__(self, sample):
+    def __call__(self, sample: Dict[str, Any]) -> Dict[str, Any]:
+        """Calls :class:`Compose`.
+
+        Parameters
+        ----------
+        sample: Dict[str, Any]
+            Dict sample.
+
+        Returns
+        -------
+        Dict[str, Any]
+            Dict sample transformed by `transforms`.
+        """
         for transform in self.transforms:
             sample = transform(sample)
 
         return sample
 
     def __repr__(self):
+        """Representation of :class:`Compose`."""
         repr_string = self.__class__.__name__ + "("
         for transform in self.transforms:
             repr_string += "\n"
@@ -50,8 +70,9 @@ class RandomFlip(DirectTransform):
     Not implemented yet.
     """
 
-    def __call__(self):
-        raise NotImplementedError
+    def __init__(self):
+        super().__init__()
+        raise NotImplementedError(":class:`RandomFlip` is not implemented yet.")
 
 
 class CreateSamplingMask(DirectModule):
@@ -60,16 +81,48 @@ class CreateSamplingMask(DirectModule):
     Creates sampling mask.
     """
 
-    def __init__(self, mask_func, shape=None, use_seed=True, return_acs=False):
+    def __init__(
+        self,
+        mask_func: Callable,
+        shape: Optional[Tuple[int, ...]] = None,
+        use_seed: bool = True,
+        return_acs: bool = False,
+    ) -> None:
+        """Inits :class:`CreateSamplingMask`.
+
+        Parameters
+        ----------
+        mask_func: Callable
+            A function which creates a sampling mask of the appropriate shape.
+        shape: tuple, optional
+            Sampling mask shape. Default: None.
+        use_seed: bool
+            If true, a pseudo-random number based on the filename is computed so that every slice of the volume get
+            the same mask every time. Default: True.
+        return_acs: bool
+            If True, it will generate an ACS mask. Default: False.
+        """
         super().__init__()
         self.mask_func = mask_func
         self.shape = shape
         self.use_seed = use_seed
         self.return_acs = return_acs
 
-    def __call__(self, sample):
+    def __call__(self, sample: Dict[str, Any]) -> Dict[str, Any]:
+        """Calls :class:`CreateSamplingMask`.
+
+        Parameters
+        ----------
+        sample: Dict[str, Any]
+            Dict sample.
+
+        Returns
+        -------
+        Dict[str, Any]
+            Sample with `sampling_mask` key.
+        """
         if not self.shape:
-            shape = sample["kspace"].shape[1:]  # ([slice], height, width, complex=2)
+            shape = sample["kspace"].shape[1:]
         elif any(_ is None for _ in self.shape):  # Allow None as values.
             kspace_shape = list(sample["kspace"].shape[1:-1])
             shape = tuple(_ if _ else kspace_shape[idx] for idx, _ in enumerate(self.shape)) + (2,)
[…]
