+++ direct/nn/didn/didn.py
@@ -33,7 +33,7 @@ class Subpixel(nn.Module):
         """
         super().__init__()
         self.conv = nn.Conv2d(
-            in_channels, out_channels * upscale_factor ** 2, kernel_size=kernel_size, padding=padding
+            in_channels, out_channels * upscale_factor**2, kernel_size=kernel_size, padding=padding
         )
         self.pixelshuffle = nn.PixelShuffle(upscale_factor)
 
