+++ direct/data/transforms.py
@@ -247,7 +247,7 @@ def modulus(data: torch.Tensor, complex_axis: int = -1) -> torch.Tensor:
 
     assert_complex(data, complex_axis)
 
-    return (data ** 2).sum(complex_axis).sqrt()  # noqa
+    return (data**2).sum(complex_axis).sqrt()  # noqa
 
 
 def roll_one_dim(data: torch.Tensor, shift: int, dim: int) -> torch.Tensor:
@@ -549,9 +549,9 @@ def root_sum_of_squares(data: torch.Tensor, dim: int = 0, complex_dim: int = -1)
 
     """
     if is_complex_data(data):
-        return torch.sqrt((data ** 2).sum(complex_dim).sum(dim))
+        return torch.sqrt((data**2).sum(complex_dim).sum(dim))
 
-    return torch.sqrt((data ** 2).sum(dim))
+    return torch.sqrt((data**2).sum(dim))
 
 
 def center_crop(data: torch.Tensor, shape: Tuple[int, int]) -> torch.Tensor:
