### 2022-04-11 bb28bbf
Add end-to-end test
