+++ direct/nn/crossdomain/multicoil.py
@@ -1,8 +1,6 @@
 # coding=utf-8
 # Copyright (c) DIRECT Contributors
 
-from typing import List
-
 import torch
 import torch.nn as nn
 
