### 2022-04-19 bceae48
Code quality (#196)
* Remove  unambiguous complex assertions (#194) (#195)
* Improve typing
* `modulus_if_complex` removed, `modulus` needs to specify axis.
* `modulus` need to specify axis
* Improve documentation and `README.rst`
* Added tests
* Add end-to-end tests
* Several bug fixes

Co-authored-by: George Yiasemis <georgeyiasemis.hotmail.com>
