+++ tests/test_train.py
@@ -1,0 +1,190 @@
+import pathlib
+import tempfile
+from dataclasses import field, make_dataclass
+
+import pytest
+from omegaconf import DictConfig, OmegaConf
+
+from direct.config.defaults import (
+    CheckpointerConfig,
+    DefaultConfig,
+    FunctionConfig,
+    InferenceConfig,
+    LossConfig,
+    ModelConfig,
+    TrainingConfig,
+    ValidationConfig,
+)
+from direct.data.datasets_config import DatasetConfig, MaskingConfig, TransformsConfig
+from direct.launch import launch
+from direct.train import setup_train
+
+
+def create_test_cfg(
+    train_dataset_shape,
+    val_dataset_shape,
+    train_batch_size,
+    val_batch_size,
+    loss_fns,
+    train_iters,
+    val_iters,
+    checkpointer_iters,
+    inference_batch_size,
+):
+    # Configs
+    transforms_config = TransformsConfig(
+        estimate_sensitivity_maps=True,
+        scaling_key="masked_kspace",
+        masking=MaskingConfig(name="FastMRIRandom"),
+        crop=(64, 64),
+    )
+
+    new_class = make_dataclass(
+        "",
+        fields=[
+            ("sample_size", int, field(init=False)),
+            ("spatial_shape", list, field(init=False)),
+            ("num_coils", int, field(init=False)),
+        ],
+        bases=(DatasetConfig,),
+    )
+
+    train_dataset_config = DatasetConfig(
+        name="FakeMRIBlobs", transforms=transforms_config, text_description="training"
+    )
+    train_dataset_config.__class__ = new_class
+    train_dataset_config.sample_size = train_dataset_shape[0]
+    train_dataset_config.num_coils = train_dataset_shape[2]
+    train_dataset_config.spatial_shape = (train_dataset_shape[1],) + train_dataset_shape[3:]
+
+    val_dataset_config = DatasetConfig(
+        name="FakeMRIBlobs", transforms=transforms_config, text_description="validation"
+    )
+    val_dataset_config.__class__ = new_class
+    val_dataset_config.sample_size = val_dataset_shape[0]
+    val_dataset_config.num_coils = val_dataset_shape[2]
+    val_dataset_config.spatial_shape = (val_dataset_shape[1],) + val_dataset_shape[3:]
+
+    checkpointer_config = CheckpointerConfig(checkpoint_steps=checkpointer_iters)
+    loss_config = LossConfig(losses=[FunctionConfig(loss) for loss in loss_fns])
+
+    training_config = TrainingConfig(
+        loss=loss_config,
+        checkpointer=checkpointer_config,
+        num_iterations=train_iters,
+        validation_steps=val_iters,
+        datasets=[train_dataset_config],
+        batch_size=train_batch_size,
+    )
+
+    validation_config = ValidationConfig(crop=None, datasets=[val_dataset_config], batch_size=val_batch_size)
+
+    inference_config = InferenceConfig(dataset=DatasetConfig(name="FakeMRIBlobs"), batch_size=inference_batch_size)
+
+    model = ModelConfig(model_name="unet.unet_2d.Unet2d")
+    config = DefaultConfig(
+        training=training_config, validation=validation_config, inference=inference_config, model=model
+    )
+    config.__class__ = make_dataclass(
+        "",
+        fields=[("additional_models", DictConfig, field(init=False))],
+        bases=(DefaultConfig,),
+    )
+    config.additional_models = DictConfig({"senistivity_model": ModelConfig(model_name="unet.unet_2d.UnetModel2d")})
+    return OmegaConf.create(config)
+
+
+@pytest.mark.parametrize(
+    "train_dataset_shape, val_dataset_shape,",
+    [[(6, 5, 3, 120, 120), (6, 5, 3, 120, 120)]],
+)
[…]
