+++ tests/tests_data/test_sens.py
@@ -1,0 +1,21 @@
+# coding=utf-8
+# Copyright (c) DIRECT Contributors
+
+import pytest
+
+from direct.data.sens import simulate_sens_maps
+
+
+@pytest.mark.parametrize(
+    "num_coils",
+    [1, 8],
+)
+@pytest.mark.parametrize(
+    "shape",
+    [(32, 32), (10, 32, 32), (11, 12, 13)],
+)
+def test_simulate_sens_maps(num_coils, shape):
+
+    sensitivity_map = simulate_sens_maps(shape, num_coils)
+
+    assert tuple(sensitivity_map.shape) == (num_coils,) + tuple(shape)
