### 2022-04-20 c7b8453
Minor edits, fake uses simulate_sensitivity_maps now
