+++ tests/tests_data/test_sens.py
@@ -3,7 +3,7 @@
 
 import pytest
 
-from direct.data.sens import simulate_sens_maps
+from direct.data.sens import simulate_sensitivity_maps
 
 
 @pytest.mark.parametrize(
@@ -14,8 +14,16 @@ from direct.data.sens import simulate_sens_maps
     "shape",
     [(32, 32), (10, 32, 32), (11, 12, 13)],
 )
-def test_simulate_sens_maps(num_coils, shape):
+@pytest.mark.parametrize(
+    "var",
+    [0.5],
+)
+@pytest.mark.parametrize(
+    "seed",
+    [None, 0],
+)
+def test_simulate_sens_maps(num_coils, shape, var, seed):
 
-    sensitivity_map = simulate_sens_maps(shape, num_coils)
+    sensitivity_map = simulate_sensitivity_maps(shape, num_coils, var, seed)
 
     assert tuple(sensitivity_map.shape) == (num_coils,) + tuple(shape)
