### 2022-04-21 be3def9
Add Shepp Logan dataset
