+++ direct/data/datasets.py
@@ -8,6 +8,7 @@ import contextlib
 import logging
 import pathlib
 import xml.etree.ElementTree as etree  # nosec
+from enum import Enum
 from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Union
 
 import numpy as np
@@ -16,6 +17,7 @@ from torch.utils.data import Dataset, IterableDataset
 
 from direct.data.fake import FakeMRIData
 from direct.data.h5_data import H5SliceData
+from direct.data.sens import simulate_sensitivity_maps
 from direct.types import PathOrString
 from direct.utils import remove_keys, str_to_class
 
@@ -590,3 +592,416 @@ def build_dataset_from_input(
         **remove_keys(dict(dataset_config), ["transforms", "lists"]),
     )
     return dataset
+
+
+class ImageIntensityMode(str, Enum):
+
+    proton = "PROTON"
+    t1 = "T1"
+    t2 = "T2"
+
+
+class SheppLoganDataset(Dataset):
+    """Shepp Logan Dataset for MRI as implemented in [1]_.
+
+    References
+    ----------
+    .. [1] Gach, H. Michael, Costin Tanase, and Fernando Boada. "2D & 3D Shepp-Logan phantom standards for MRI." 2008 19th International Conference on Systems Engineering. IEEE, 2008.
+    """
+
+    GYROMAGNETIC_RATIO: float = 267.52219
+    DEFAULT_NUM_ELLIPSOIDS: int = 15
+    ELLIPSOID_NUM_PARAMS: int = 13
+    IMAGE_INTENSITIES: List[str] = ["PROTON", "T1", "T2"]
+
+    def __init__(
+        self,
+        shape: Union[int, Union[List[int], Tuple[int, int, int]]],
+        num_coils: int,
+        intensity: ImageIntensityMode,
+        seed: Optional[Union[int, List[int]]] = None,
+        ellipsoids: np.ndarray = None,
+        B0: float = 3.0,
+        T2_star: Optional[bool] = None,
+        zlimits: Tuple[int, int] = (-1, 1),
+        transform: Optional[Callable] = None,
+        text_description: Optional[str] = None,
+    ) -> None:
+        r"""Inits :class:`SheppLoganDataset`.
+
+        Parameters
+        ----------
+        shape: Union[int, Union[List[int], Tuple[int, int, int]]]
+            Shape of Shepp Logan phantom (3-dimensional).
+        num_coils: int
+            Number of simulated coils.
+        intensity: ImageIntensityMode
+            Can be `PROTON` to return the proton density dataset, `T1` or `T2`.
+        seed: Optional[Union[int, List[int]]]
+            Seed to be used for coil sensitivity maps. Default: None.
+        ellipsoids: np.ndarray
+            Ellipsoids parameters. If None, it will used the default parameters as per the paper. Default: None.
+        B0: float
+            Magnetic field. Default: 3.0.
+        T2_star: Optional[bool]
+            If True, a T2^{*} dataset will be output. Only valid for intensity = `T2`. Default: None.
+        zlimits: Tuple[int, int]
+            Limits of z-axis. Default: (-1, 1).
+        transform: Optional[Callable]
+            A list of transforms to be applied on the generated samples. Default is None.
+        text_description: Optional[str]
+            Description of dataset, can be useful for logging. Default: None.
+        """
+        self.logger = logging.getLogger(type(self).__name__)
+
+        (self.nx, self.ny, self.nz) = (shape, shape, shape) if isinstance(shape, int) else tuple(shape)
+        self.num_coils = num_coils
+
[…]
