### 2022-04-22 c149101
fixes codacy errors
