### 2022-04-23 f1b4169
Update colab.rst
