### 2022-04-26 bbc2f16
Add CC example
