+++ docs/examples.rst
@@ -1,0 +1,1 @@
+.. include:: ./examples/calgary_campinas_dataset.rst
