### 2022-04-26 381f50c
Add examples (#199)
Adds:
* How to run on colab example
* How to set up the Calgary-Campinas dataset example
