+++ docs/index.rst
@@ -38,6 +38,13 @@ images such as MRIs from partially observed or noisy input data.
 
 .. toctree::
    :maxdepth: 1
+   :caption: Examples
+
+   examples.rst
+   colab.rst
+
+.. toctree::
+   :maxdepth: 1
    :caption: Model Zoo
 
    model_zoo.rst
