### 2022-04-28 4e9e99f
Add kspace cropping transform
