+++ direct/data/mri_transforms.py
@@ -157,30 +157,68 @@ class CreateSamplingMask(DirectModule):
         return sample
 
 
-class CropAndMask(DirectModule):
+class ApplyMask(DirectModule):
     """Data Transformer for training MRI reconstruction models.
 
-    Crops and masks k-space using a sampling mask.
+    Masks the k-space using a sampling mask.
+    """
+
+    def __init__(self) -> None:
+        """Inits :class:`ApplyMask`."""
+        super().__init__()
+        self.logger = logging.getLogger(type(self).__name__)
+
+    def __call__(self, sample: Dict[str, Any]) -> Dict[str, Any]:
+        """Calls :class:`ApplyMask`.
+
+        This assumes that a `sampling_mask` is present in the sample.
+
+        Parameters
+        ----------
+        sample: Dict[str, Any]
+            Dict sample containing key `kspace`.
+
+        Returns
+        -------
+        Dict[str, Any]
+            Sample with new key `masked_kspace`.
+        """
+        kspace = sample["kspace"]
+
+        assert "sampling_mask" in sample, "Key 'sampling_mask' not found in sample."
+        sampling_mask = sample["sampling_mask"]
+
+        masked_kspace, _ = T.apply_mask(kspace, sampling_mask)
+        sample["masked_kspace"] = masked_kspace
+
+        return sample
+
+
+class CropKspace(DirectModule):
+    """Data Transformer for training MRI reconstruction models.
+
+    Crops the k-space by:
+        * It first projects the k-space to the image-domain via the backward operator,
+        * It crops the back-projected k-space to specified shape or key,
+        * It transforms the cropped back-projected k-space to the k-space domain via the forward operator.
     """
 
     def __init__(
         self,
-        crop: Union[None, Tuple[int, ...]],
-        use_seed: bool = True,
+        crop: Union[str, Tuple[int, ...], List[int]],
         forward_operator: Callable = T.fft2,
         backward_operator: Callable = T.ifft2,
         image_space_center_crop: bool = False,
         random_crop_sampler_type: Optional[str] = "uniform",
+        random_crop_sampler_use_seed: Optional[bool] = True,
+        random_crop_sampler_gaussian_sigma: Optional[List[float]] = None,
     ) -> None:
-        """Inits :class:`CropAndMask`.
+        """Inits :class:`CropKspace`.
 
         Parameters
         ----------
-        crop: tuple of ints or None
-            Size to crop input_image to.
-        use_seed: bool
-            If true, a pseudo-random number based on the filename is computed so that every slice of the volume get
-            the same mask every time. Default: True.
+        crop: tuple of ints or str
+            Shape to crop the input to or a string pointing to a crop key (e.g. `reconstruction_size`).
         forward_operator: Callable
             The forward operator, e.g. some form of FFT (centered or uncentered).
             Default: :class:`direct.data.transforms.fft2`.
@@ -191,87 +229,73 @@ class CropAndMask(DirectModule):
             If set, the crop in the data will be taken in the center
         random_crop_sampler_type: Optional[str]
             If "uniform" the random cropping will be done by uniformly sampling `crop`, as opposed to `gaussian` which
-            will sample from a gaussian distribution. Default: "uniform".
+            will sample from a gaussian distribution. If `image_space_center_crop` is True, then this is ignored.
+            Default: "uniform".
+        random_crop_sampler_use_seed: bool
+            If true, a pseudo-random number based on the filename is computed so that every slice of the volume
+            is cropped the same way. Default: True.
+        random_crop_sampler_gaussian_sigma: Optional[List[float]]
+            Standard variance of the gaussian when `random_crop_sampler_type` is `gaussian`.
[…]

[…]
