### 2022-05-03 NKI-AI/direct#191 @Winnie-xwj
I have used the Conda instructions(https://docs.aiforoncology.nl/direct/installation.html#install-using-conda) and worker with Colab, but I have some problems in reproducing the experiment in their paper (Recurrent Variational Network: A Deep Learning Inverse Problem Solver applied to the task of Accelerated MRI Reconstruction (CVPR 2022)[¶](https://docs.aiforoncology.nl/direct/cvpr2022_recurrentvarnet.html#recurrent-variational-network-a-deep-learning-inverse-problem-solver-applied-to-the-task-of-accelerated-mri-reconstruction-cvpr-2022)
Could you give me some advice?
