### 2022-05-04 6a5bca3
Add datasets configs, minor SheppLogan datasets changes,  and  now work with **kwargs, training_root and validation_root not required now for train - some datasets don't require saved data.
