+++ direct/data/datasets.py
@@ -2,11 +2,11 @@
 # Copyright (c) DIRECT Contributors
 
 """DIRECT datasets module."""
-
 import bisect
 import contextlib
 import logging
 import pathlib
+import sys
 import xml.etree.ElementTree as etree  # nosec
 from enum import Enum
 from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Union
@@ -22,6 +22,17 @@ from direct.types import PathOrString
 from direct.utils import remove_keys, str_to_class
 
 logger = logging.getLogger(__name__)
+
+__all__ = [
+    "build_dataset_from_input",
+    "CalgaryCampinasDataset",
+    "ConcatDataset",
+    "FastMRIDataset",
+    "FakeMRIBlobsDataset",
+    "SheppLoganT1Dataset",
+    "SheppLoganT2Dataset",
+    "SheppLoganProtonDataset",
+]
 
 
 @contextlib.contextmanager
@@ -116,7 +127,7 @@ class FakeMRIBlobsDataset(Dataset):
 
         if len(spatial_shape) not in [2, 3]:
             raise NotImplementedError(
-                f"Currently FakeDataset is implemented only for 2D or 3D data."
+                f"Currently FakeDataset is implemented only for 2D or 3D data. "
                 f"Spatial shape must have 2 or 3 dimensions. Got shape {spatial_shape}."
             )
         self.sample_size = sample_size
@@ -260,9 +271,11 @@ class FastMRIDataset(H5SliceData):
 
     def __init__(
         self,
-        root: pathlib.Path,
+        data_root: pathlib.Path,
         transform: Optional[Callable] = None,
         filenames_filter: Optional[List[PathOrString]] = None,
+        filenames_lists: Union[List[PathOrString], None] = None,
+        filenames_lists_root: Union[PathOrString, None] = None,
         regex_filter: Optional[str] = None,
         pass_mask: bool = False,
         pass_max: bool = True,
@@ -280,8 +293,10 @@ class FastMRIDataset(H5SliceData):
         extra_keys.append("ismrmrd_header")
 
         super().__init__(
-            root=root,
+            root=data_root,
             filenames_filter=filenames_filter,
+            filenames_lists=filenames_lists,
+            filenames_lists_root=filenames_lists_root,
             regex_filter=regex_filter,
             metadata=None,
             extra_keys=tuple(extra_keys),
@@ -384,18 +399,22 @@ class CalgaryCampinasDataset(H5SliceData):
 
     def __init__(
         self,
-        root: pathlib.Path,
+        data_root: pathlib.Path,
         transform: Optional[Callable] = None,
         regex_filter: Optional[str] = None,
         filenames_filter: Optional[List[PathOrString]] = None,
+        filenames_lists: Union[List[PathOrString], None] = None,
+        filenames_lists_root: Union[PathOrString, None] = None,
         pass_mask: bool = False,
         crop_outer_slices: bool = False,
         pass_h5s: Optional[Dict] = None,
         **kwargs,
     ) -> None:
         super().__init__(
-            root=root,
+            root=data_root,
             filenames_filter=filenames_filter,
+            filenames_lists=filenames_lists,
+            filenames_lists_root=filenames_lists_root,
             regex_filter=regex_filter,
             metadata=None,
             extra_keys=None,
@@ -442,13 +461,12 @@ class ConcatDataset(Dataset):
     """Dataset as a concatenation of multiple datasets.
 
     This class is useful to assemble different existing datasets.
+    From pytorch 1.5.1: :class:`torch.utils.data.ConcatDataset`.
 
     Parameters
     ----------
     datasets: sequence
         List of datasets to be concatenated
-
[…]
