+++ tests/tests_utils/test_utils.py
@@ -12,7 +12,7 @@ import torch
 from direct.utils import is_power_of_two, normalize_image, remove_keys, set_all_seeds
 from direct.utils.asserts import assert_complex
 from direct.utils.bbox import crop_to_largest
-from direct.utils.dataset import get_filenames_for_datasets
+from direct.utils.dataset import get_filenames_for_datasets_from_config
 
 
 def create_input(shape):
@@ -95,7 +95,7 @@ def test_get_filenames_for_datasets(file_list, num_samples):
                 f.write(f"file_{_}.txt" + "\n")
 
         cfg = mock_cfg(lists=["mock_list.lst"]) if file_list else mock_cfg()
-        filenames = get_filenames_for_datasets(cfg, files_root=path_to_list, data_root=data_root)
+        filenames = get_filenames_for_datasets_from_config(cfg, files_root=path_to_list, data_root=data_root)
         if file_list:
             assert len(filenames) == num_samples
         else:
