### 2022-05-04 80f9a51
now doesn't require --data-root arg, edit dataset config attributes to allow for file filtering, minor fixes
