### 2022-05-05 e078276
Update docs to comply with new train/predict changes
