+++ docs/inference.rst
@@ -10,28 +10,37 @@ To perform inference on a single machine run the following code block in your li
 
 .. code-block:: bash
 
-    $ direct predict <data_root> <output_directory> --cfg <cfg_path_or_url> --checkpoint <checkpoint_path_or_url> --num-gpus <num_gpus> [ --cfg <cfg_filename>.yaml --other-flags <other_flags>]
-                  
+    $ direct predict <output_directory> --cfg <cfg_path_or_url> --checkpoint <checkpoint_path_or_url> \
+            --num-gpus <num_gpus> [--data-root <data_root>] [ --cfg <cfg_filename>.yaml --other-flags <other_flags>]
+
 To predict using multiple machines run the following code (one command on each machine):
 
 .. code-block:: bash
 
-    (machine0)$ direct predict <data_root> <output_directory> --cfg <cfg_path_or_url> --checkpoint <checkpoint_path_or_url> --machine-rank 0 --num-machines 2 --dist-url <URL> [--other-flags]
-    (machine1)$ direct predict <data_root> <output_directory> --cfg <cfg_path_or_url> --checkpoint <checkpoint_path_or_url> --machine-rank 1 --num-machines 2 --dist-url <URL> [--other-flags]
+    (machine0)$ direct predict <output_directory> --cfg <cfg_path_or_url> --checkpoint <checkpoint_path_or_url> \
+            --machine-rank 0 --num-machines 2 --dist-url <URL> [--data-root <data_root>] [--other-flags]
+    (machine1)$ direct predict <output_directory> --cfg <cfg_path_or_url> --checkpoint <checkpoint_path_or_url> \
+            --machine-rank 1 --num-machines 2 --dist-url <URL> [--data-root <data_root>] [--other-flags]
 
-The ``cfg_path_or_url`` should point to a configuration file that includes all the model parameters used for the trained model checkpoint ``checkpoint_path_or_url`` and should also include an inference configuration as follows:
+The ``cfg_path_or_url`` should point to a configuration file that includes all the model parameters used for the trained
+model checkpoint ``checkpoint_path_or_url`` and should also include an inference configuration as follows:
 
 .. code-block:: yaml
 
     inference:
-  dataset:
-    name: InferenceDataset
-    lists: ...
-    transforms:
-      masking:
+      dataset:
+        name: InferenceDataset
+        <lists: ...>
+        transforms:
+          masking:
+            ...
+          ...
+        text_description: <inference_description>
         ...
-      ...
-    text_description: <inference_description>
-    ...
-  batch_size: <batch_size>
+      batch_size: <batch_size>
   ...
+
+Notes
+-----
+Note that if the `dataset` you want to use requires downloaded path, you can pass the `--data-root` argument which
+points to the data directory.
