### 2022-05-05 ef44f8c
Update calgary campinas projects to comply with new train/predict changes
