### 2022-05-09 fa3b163
Use inference.crop and inference.batch_size in direct predict
