### 2022-05-09 f58305e
Update calgary campinas readme
