### 2022-05-09 402812a
Add usage example for shepplogan dataset
