+++ docs/index.rst
@@ -49,6 +49,7 @@ images such as MRIs from partially observed or noisy input data.
    calgary_campinas
    spie2022_radial_subsampling
    cvpr2022_recurrentvarnet
+   shepp_logan
 
 .. toctree::
    :maxdepth: 2
