+++ direct/inference.py
@@ -4,13 +4,15 @@
 import logging
 import sys
 from functools import partial
-from typing import Optional
+from typing import Callable, DefaultDict, Dict, List, Optional, Union
 
 import torch
+from omegaconf import DictConfig
 
 from direct.data.datasets import build_dataset_from_input
 from direct.data.mri_transforms import build_mri_transforms
 from direct.environment import setup_inference_environment
+from direct.types import FileOrUrl, PathOrString
 from direct.utils import chunks, remove_keys
 from direct.utils.io import read_list
 from direct.utils.writers import write_output_to_h5
@@ -19,39 +21,60 @@ logger = logging.getLogger(__name__)
 
 
 def setup_inference_save_to_h5(
-    get_inference_settings,
-    run_name,
-    data_root,
-    base_directory,
-    output_directory,
-    filenames_filter,
-    checkpoint,
-    device,
+    get_inference_settings: Callable,
+    run_name: str,
+    data_root: Union[PathOrString, None],
+    base_directory: PathOrString,
+    output_directory: PathOrString,
+    filenames_filter: Union[List[PathOrString], None],
+    checkpoint: FileOrUrl,
+    device: str,
     num_workers: int,
     machine_rank: int,
-    cfg_file=None,
+    cfg_file: Union[PathOrString, None] = None,
     process_per_chunk: Optional[int] = None,
     mixed_precision: bool = False,
     debug: bool = False,
-):
+    is_validation: bool = False,
+) -> None:
     """This function contains most of the logic in DIRECT required to launch a multi-gpu / multi-node inference process.
+
+    It saves predictions as `.h5` files.
 
     Parameters
     ----------
     get_inference_settings: Callable
-    run_name:
-    data_root:
-    base_directory:
-    output_directory:
-    filenames_filter:
-    checkpoint:
-    device:
-    num_workers:
-    machine_rank:
-    cfg_file:
-    process_per_chunk:
-    mixed_precision:
-    debug:
+        Callable object to create inference dataset and environment.
+    run_name: str
+        Experiment run name. Can be an empty string.
+    data_root: Union[PathOrString, None]
+        Path of the directory of the data if applicable for dataset. Can be None.
+    base_directory: PathOrString
+        Path to directory where where inference logs will be stored. If `run_name` is not an empty string,
+        `base_directory / run_name` will be used.
+    output_directory: PathOrString
+        Path to directory where output data will be saved.
+    filenames_filter: Union[List[PathOrString], None]
+        List of filenames to include in the dataset (if applicable). Can be None.
+    checkpoint: FileOrUrl
+        Checkpoint to a model. This can be a path to a local file or an URL.
+    device: str
+        Device name.
+    num_workers: int
+        Number of workers.
+    machine_rank: int
+        Machine rank.
+    cfg_file: Union[PathOrString, None]
+        Path to configuration file. If None, will search in `base_directory`.
+    process_per_chunk: int
+        Processes per chunk number.
+    mixed_precision: bool
+        If True, mixed precision will be allowed. Default: False.
+    debug: bool
+        If True, debug information will be displayed. Default: False.
+    is_validation: bool
+        If True, will use settings (e.g. `batch_size` & `crop`) of `validation` in config.
+        Otherwise it will use `inference` settings. Default: False.
 
     Returns
     -------
@@ -65,9 +88,9 @@ def setup_inference_save_to_h5(
 
     # Trigger cudnn benchmark when the number of different input masks_dict is small.
     torch.backends.cudnn.benchmark = True
-
-    if filenames_filter:
[…]
