+++ direct/train.py
@@ -23,7 +23,7 @@ from direct.environment import setup_training_environment
 from direct.launch import launch
 from direct.types import PathOrString
 from direct.utils import remove_keys, set_all_seeds, str_to_class
-from direct.utils.dataset import get_filenames_for_datasets
+from direct.utils.dataset import get_filenames_for_datasets_from_config
 from direct.utils.io import check_is_valid_url, read_json
 
 logger = logging.getLogger(__name__)
@@ -87,33 +87,34 @@ def build_transforms_from_environment(env, dataset_config: DictConfig) -> Callab
 def build_training_datasets_from_environment(
     env,
     datasets_config: List[DictConfig],
-    lists_root: PathOrString,
-    data_root: PathOrString,
-    initial_images: Union[List[pathlib.Path], None] = None,
-    initial_kspaces: Union[List[pathlib.Path], None] = None,
+    lists_root: Optional[PathOrString] = None,
+    data_root: Optional[PathOrString] = None,
+    initial_images: Optional[Union[List[pathlib.Path], None]] = None,
+    initial_kspaces: Optional[Union[List[pathlib.Path], None]] = None,
     pass_text_description: bool = True,
     pass_dictionaries: Optional[Dict[str, Dict]] = None,
-    **kwargs,
 ):
 
     datasets = []
     for idx, dataset_config in enumerate(datasets_config):
         if pass_text_description:
-            if not dataset_config.text_description:
+            if not "text_description" in dataset_config:
                 dataset_config.text_description = f"ds{idx}" if len(datasets_config) > 1 else None
         else:
             dataset_config.text_description = None
         transforms = build_transforms_from_environment(env, dataset_config)
-        filenames_filter = get_filenames_for_datasets(dataset_config, lists_root, data_root)
-        dataset = build_dataset_from_input(
-            transforms,
-            dataset_config,
-            initial_images,
-            initial_kspaces,
-            filenames_filter,
-            data_root,
-            pass_dictionaries,
-        )
+        dataset_args = {"transforms": transforms, "dataset_config": dataset_config}
+        if initial_images is not None:
+            dataset_args.update({"initial_images": initial_images})
+        if initial_kspaces is not None:
+            dataset_args.update({"initial_kspaces": initial_kspaces})
+        if data_root is not None:
+            dataset_args.update({"data_root": data_root})
+            filenames_filter = get_filenames_for_datasets_from_config(dataset_config, lists_root, data_root)
+            dataset_args.update({"filenames_filter": filenames_filter})
+        if pass_dictionaries is not None:
+            dataset_args.update({"pass_dictionaries": pass_dictionaries})
+        dataset = build_dataset_from_input(**dataset_args)
 
         logger.debug(f"Transforms {idx + 1} / {len(datasets_config)} :\n{transforms}")
         datasets.append(dataset)
@@ -126,22 +127,22 @@ def build_training_datasets_from_environment(
 
 
 def setup_train(
-    run_name,
-    training_root,
-    validation_root,
-    base_directory,
-    cfg_filename,
-    force_validation,
-    initialization_checkpoint,
-    initial_images,
-    initial_kspace,
-    noise,
-    device,
-    num_workers,
-    resume,
-    machine_rank,
-    mixed_precision,
-    debug,
+    run_name: str,
+    training_root: Union[pathlib.Path, None],
+    validation_root: Union[pathlib.Path, None],
+    base_directory: pathlib.Path,
+    cfg_filename: PathOrString,
+    force_validation: bool,
+    initialization_checkpoint: PathOrString,
+    initial_images: Optional[Union[List[pathlib.Path], None]],
[…]
