+++ direct/utils/dataset.py
@@ -24,9 +24,9 @@ def get_filenames_for_datasets_from_config(cfg, files_root: PathOrString, data_r
     """
     if "filenames_lists" not in cfg:
         return None
-    else:
-        lists = cfg.filenames_lists
-        return get_filenames_for_datasets(lists, files_root, data_root)
+
+    lists = cfg.filenames_lists
+    return get_filenames_for_datasets(lists, files_root, data_root)
 
 
 def get_filenames_for_datasets(lists: List[PathOrString], files_root: PathOrString, data_root: pathlib.Path):
