+++ direct/cli/utils.py
@@ -19,11 +19,12 @@ def is_file(path):
 def file_or_url(path: PathOrString) -> FileOrUrl:
     if check_is_valid_url(path):
         return FileOrUrl(path)
-    else:
-        path = pathlib.Path(path)
-        if path.is_file():
-            return FileOrUrl(path)
-        raise argparse.ArgumentTypeError(f"{path} is not a valid file or url.")
+
+    path = pathlib.Path(path)
+    if path.is_file():
+        return FileOrUrl(path)
+
+    raise argparse.ArgumentTypeError(f"{path} is not a valid file or url.")
 
 
 def check_train_val(key, name):
