### 2022-05-16 #4278 @georgeyiasemis
> @georgeyiasemis Great job on the package! I thoroughly enjoyed playing with it for a while. However, I could not get your docker to work.Can you please close the [Docker issue](https://github.com/NKI-AI/direct/issues/191) that @estenhl and I both reported? I will then check off the remaining items.

Thanks a lot @sairamgeethanath 👍 ! Yes we will do that!
