+++ docker/Dockerfile
@@ -1,10 +1,9 @@
-ARG CUDA="10.2"
-ARG CUDNN="8"
-ARG PYTORCH="1.7"
+ARG CUDA="11.3.0"
+ARG PYTORCH="1.10"
 ARG PYTHON="3.8"
 
 # TODO: conda installs its own version of cuda
-FROM nvidia/cuda:${CUDA}-cudnn${CUDNN}-devel-ubuntu18.04
+FROM nvidia/cuda:${CUDA}-devel-ubuntu18.04
 
 ENV CUDA_PATH /usr/local/cuda
 ENV CUDA_ROOT /usr/local/cuda/bin
@@ -17,10 +16,10 @@ RUN apt-get -qq update && apt-get install -y --no-install-recommends libxext6 li
     && rm -rf /var/lib/apt/lists/* \
     && ldconfig
 
-RUN git clone https://github.com/mrirecon/bart.git /tmp/bart
-    && cd /tmp/bart
-    && make -j4
-    && make install
+RUN git clone https://github.com/mrirecon/bart.git /tmp/bart \
+    && cd /tmp/bart \
+    && make -j4 \
+    && make install \
     && rm -rf /tmp/bart
 
 # Make a user
@@ -41,27 +40,29 @@ ENV PATH "/users/direct/miniconda3/bin:/tmp/bart/:$PATH:$CUDA_ROOT"
 # Setup python packages
 RUN conda update -n base conda -yq \
  && conda install python=${PYTHON} \
+ && conda install jupyter \
  && conda install cudatoolkit=${CUDA} torchvision -c pytorch
 
 RUN if [ "nightly$PYTORCH" = "nightly" ] ; then echo "Installing pytorch nightly" && \
     conda install pytorch -c pytorch-nightly; else conda install pytorch=${PYTORCH} -c pytorch ; fi
 
 USER root
-RUN mkdir /direct && chmod 777 /direct
+RUN mkdir direct:direct /direct && chown direct:direct /direct && chmod 777 /direct
 
 USER direct
 
 RUN jupyter notebook --generate-config
 ENV CONFIG_PATH "/users/direct/.jupyter/jupyter_notebook_config.py"
-COPY "jupyter_notebook_config.py" ${CONFIG_PATH}
+COPY "docker/jupyter_notebook_config.py" ${CONFIG_PATH}
 
 # Copy files into the docker
 COPY [".", "/direct"]
 WORKDIR /direct
+USER root
 RUN python -m pip install -e ".[dev]"
+USER direct
 
 ENV PYTHONPATH /tmp/bart/python:/direct
-WORKDIR /direct
 
 # Provide an open entrypoint for the docker
 ENTRYPOINT $0 $@
