+++ docker/README.rst
@@ -1,15 +1,15 @@
 Docker Installation
 -------------------
 
-Use the container (docker ≥ 19.03 required)
--------------------------------------------
+Use the container
+~~~~~~~~~~~~~~~~~
 
 To build:
 
 .. code-block:: bash
 
-    cd docker/
-    docker build -t direct:latest .
+    cd direct/
+    docker build -t direct:latest -f docker/Dockerfile .
 
 To run using all GPUs:
 
@@ -18,3 +18,8 @@ To run using all GPUs:
     docker run --gpus all -it \
         --shm-size=24gb --volume=<source_to_data>:/data --volume=<source_to_results>:/output \
 	    --name=direct direct:latest /bin/bash
+
+Requirements
+~~~~~~~~~~~~
+
+* docker ≥ 19.03
