### 2022-05-16 99a288e
Update docker image (#204), closes #191
