### 2022-05-16 #4278 @sairamgeethanath
I will test it by the end of the week.

Best,
Sairam

On Mon, May 16, 2022 at 1:53 PM George Yiasemis ***@***.***>
wrote:

> @georgeyiasemis <https://github.com/georgeyiasemis> Great job on the
> package! I thoroughly enjoyed playing with it for a while. However, I could
> not get your docker to work.Can you please close the Docker issue
> <https://github.com/NKI-AI/direct/issues/191> that @estenhl
> <https://github.com/estenhl> and I both reported? I will then check off
> the remaining items.
>
> @sairamgeethanath <https://github.com/sairamgeethanath> The docker issue
> is fixed!
>
