### 2022-05-20 5826d15
Shepp-Logan phantom & sensitivity map simulator (#202)
* Add Shepp-Logan Datasets
* Add sensitivity map simulator
* Refactor `direct train`, `direct predict` and `python3 projects/predict_val.py` to not necessarily require path to data as some datasets don't require it - `build_dataset_from_input` relies on `**kwargs` now
* Refactored configs and docs to comply with the above
