+++ direct/data/datasets.py
@@ -2,12 +2,13 @@
 # Copyright (c) DIRECT Contributors
 
 """DIRECT datasets module."""
-
 import bisect
 import contextlib
 import logging
 import pathlib
+import sys
 import xml.etree.ElementTree as etree  # nosec
+from enum import Enum
 from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Union
 
 import numpy as np
@@ -16,10 +17,23 @@ from torch.utils.data import Dataset, IterableDataset
 
 from direct.data.fake import FakeMRIData
 from direct.data.h5_data import H5SliceData
+from direct.data.sens import simulate_sensitivity_maps
 from direct.types import PathOrString
 from direct.utils import remove_keys, str_to_class
 
 logger = logging.getLogger(__name__)
+
+__all__ = [
+    "build_dataset_from_input",
+    "CalgaryCampinasDataset",
+    "ConcatDataset",
+    "FastMRIDataset",
+    "FakeMRIBlobsDataset",
+    "SheppLoganDataset",
+    "SheppLoganT1Dataset",
+    "SheppLoganT2Dataset",
+    "SheppLoganProtonDataset",
+]
 
 
 @contextlib.contextmanager
@@ -114,7 +128,7 @@ class FakeMRIBlobsDataset(Dataset):
 
         if len(spatial_shape) not in [2, 3]:
             raise NotImplementedError(
-                f"Currently FakeDataset is implemented only for 2D or 3D data."
+                f"Currently FakeDataset is implemented only for 2D or 3D data. "
                 f"Spatial shape must have 2 or 3 dimensions. Got shape {spatial_shape}."
             )
         self.sample_size = sample_size
@@ -124,7 +138,7 @@ class FakeMRIBlobsDataset(Dataset):
         self.pass_attrs = pass_attrs if pass_attrs is not None else True
         self.text_description = text_description
         if self.text_description:
-            self.logger.info(f"Dataset description: {self.text_description}.")
+            self.logger.info("Dataset description: %s.", self.text_description)
 
         self.fake_data: Callable = FakeMRIData(
             ndim=len(self.spatial_shape),
@@ -164,7 +178,8 @@ class FakeMRIBlobsDataset(Dataset):
         current_slice_number = 0
         for idx, filename in enumerate(filenames):
             if len(filenames) < 5 or idx % (len(filenames) // 5) == 0 or len(filenames) == (idx + 1):
-                self.logger.info("Parsing: {(idx + 1) / len(filenames) * 100:.2f}%.")
+                # pylint: disable=logging-fstring-interpolation
+                self.logger.info(f"Parsing: {(idx + 1) / len(filenames) * 100:.2f}%.")
 
             num_slices = self.spatial_shape[0] if len(self.spatial_shape) == 3 else 1
             self.volume_indices[pathlib.PosixPath(filename)] = range(
@@ -258,9 +273,11 @@ class FastMRIDataset(H5SliceData):
 
     def __init__(
         self,
-        root: pathlib.Path,
+        data_root: pathlib.Path,
         transform: Optional[Callable] = None,
         filenames_filter: Optional[List[PathOrString]] = None,
+        filenames_lists: Union[List[PathOrString], None] = None,
+        filenames_lists_root: Union[PathOrString, None] = None,
         regex_filter: Optional[str] = None,
         pass_mask: bool = False,
         pass_max: bool = True,
@@ -278,8 +295,10 @@ class FastMRIDataset(H5SliceData):
         extra_keys.append("ismrmrd_header")
 
[…]
