+++ direct/data/fake.py
@@ -1,11 +1,14 @@
 # coding=utf-8
 # Copyright (c) DIRECT Contributors
+
 import logging
 import pathlib
 from typing import Dict, List, Optional, Tuple, Union
 
 import numpy as np
 from sklearn.datasets import make_blobs
+
+from direct.data.sens import simulate_sensitivity_maps
 
 logger = logging.getLogger(__name__)
 
@@ -19,13 +22,16 @@ class FakeMRIData:
         blobs_n_samples: Optional[int] = None,
         blobs_cluster_std: Optional[float] = None,
     ) -> None:
-        """
+        """Inits :class:`FakeMRIData`.
 
         Parameters
         ----------
         ndim: int
-        blobs_n_samples: Optional[int], default is None.
-        blobs_cluster_std: Optional[float], default is None.
+            Dimension of samples. Can be 2 or 3. Default: 2.
+        blobs_n_samples: Optional[int]
+            The total number of points equally divided among clusters. Default: None.
+        blobs_cluster_std: Optional[float]
+            Standard deviation of the clusters. Default: None.
         """
 
         if ndim not in [2, 3]:
@@ -49,12 +55,14 @@ class FakeMRIData:
         num_coils: int
         """
 
-        samples, centers, classes = self.make_blobs(spatial_shape, num_coils)
+        samples = self.make_blobs(spatial_shape, num_coils)
 
-        image = self._get_image_from_samples(samples, classes, num_coils, spatial_shape)
-
+        image = self._get_image_from_samples(samples, spatial_shape)
+        image = image[None]
         if num_coils > 1:
-            image = self._make_coil_data(image, samples, centers, classes)
+            sens_maps = simulate_sensitivity_maps(spatial_shape[-2:], num_coils)
+
+            image = image * (sens_maps if self.ndim == 2 else sens_maps[:, None])
 
         kspace = fft(image)
 
@@ -79,14 +87,14 @@ class FakeMRIData:
     ) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
         """Generates gaussian blobs in 'num_coils' classes and scales them the interval.
 
-        [0, slice] x [0, heihgt] x [0, width].
+        [0, slice] x [0, height] x [0, width].
         """
 
         # Number of samples to be converted to an image
         n_samples = self.blobs_n_samples if self.blobs_n_samples else np.prod(list(spatial_shape)) // self.ndim
         cluster_std = self.blobs_cluster_std if self.blobs_cluster_std is not None else 0.1
 
-        samples, classes, centers = make_blobs(
+        samples, _, _ = make_blobs(
             n_samples=n_samples,
             n_features=self.ndim,
             centers=num_coils,
@@ -96,55 +104,16 @@ class FakeMRIData:
             return_centers=True,
         )
 
-        scaled_samples = scale_data(data=samples, shape=spatial_shape)
-        scaled_centers = scale_data(data=centers, other=samples, shape=spatial_shape)
+        samples = scale_data(data=samples, shape=spatial_shape)
 
-        return scaled_samples, scaled_centers, classes
-
-    def _get_image_from_samples(self, samples, classes, num_coils, spatial_shape):
-        image = np.zeros(tuple([num_coils] + list(spatial_shape)))
-        for coil_idx in range(num_coils):
-
-            if self.ndim == 2:
-                image[
-                    coil_idx, samples[np.where(classes == coil_idx), 0], samples[np.where(classes == coil_idx), 1]
-                ] = 1  # assign 1 to each pixel
-
-            elif self.ndim == 3:
-                image[
-                    coil_idx,
-                    samples[np.where(classes == coil_idx), 0],
-                    samples[np.where(classes == coil_idx), 1],
[…]
