+++ direct/cli/utils.py
@@ -1,9 +1,11 @@
 # coding=utf-8
 # Copyright (c) DIRECT Contributors
+
 import argparse
 import pathlib
 import sys
 
+from direct.types import FileOrUrl, PathOrString
 from direct.utils.io import check_is_valid_url
 
 
@@ -14,12 +16,14 @@ def is_file(path):
     raise argparse.ArgumentTypeError(f"{path} is not a valid file or url.")
 
 
-def file_or_url(path):
+def file_or_url(path: PathOrString) -> FileOrUrl:
     if check_is_valid_url(path):
-        return path
+        return FileOrUrl(path)
+
     path = pathlib.Path(path)
     if path.is_file():
-        return path
+        return FileOrUrl(path)
+
     raise argparse.ArgumentTypeError(f"{path} is not a valid file or url.")
 
 
