+++ direct/__init__.py
@@ -2,4 +2,4 @@
 # Copyright (c) DIRECT Contributors
 
 __author__ = """direct contributors"""
-__version__ = "1.0.1"
+__version__ = "1.0.2-dev0"
