+++ paper.bib
@@ -1,11 +1,18 @@
-@misc{yiasemis2021deep,
-      title={Deep MRI Reconstruction with Radial Subsampling},
-      author={George Yiasemis and Chaoping Zhang and Clara I. Sánchez and Jan-Jakob Sonke and Jonas Teuwen},
-      year={2021a},
-      eprint={2108.07619},
-      archivePrefix={arXiv},
-      primaryClass={eess.IV}
+@inproceedings{yiasemis2021deep,
+author = {George Yiasemis and Chaoping Zhang and Clara I. Sánchez and Jan-Jakob Sonke and Jonas Teuwen},
+title = {{Deep MRI reconstruction with radial subsampling}},
+volume = {12031},
+booktitle = {Medical Imaging 2022: Physics of Medical Imaging},
+editor = {Wei Zhao and Lifeng Yu},
+organization = {International Society for Optics and Photonics},
+publisher = {SPIE},
+pages = {801 -- 810},
+keywords = {Deep MRI Reconstruction, Radial Subsampling, Subsampling Masks, Retrospective Subsampling, Recurrent Inference Machines},
+year = {2022},
+doi = {10.1117/12.2609876},
+URL = {https://doi.org/10.1117/12.2609876}
 }
+
 
 @article{fastmri2021,
    title={Results of the 2020 fastMRI Challenge for Machine Learning MR Image Reconstruction},
