### 2022-05-25 NKI-AI/direct#205 @osorensen
Suggested edits of JOSS submission
Here are some editorial suggestions for the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4278).

- On lines 7-8, please change "Imaging Processing" to "Image Processing".
- Equation 2, right after line 28, wouldn't it be correct to change `\mathcal{J}(z) ` to `\mathcal{J}(\vec{z})`?
- Line 41, change "algorithmsp" to "algorithms".
- Line 60, I suggest rewriting "In a configuration file can be specified all the experiment parameters." to "All experiment parameters can be specified in a configuration file."
