### 2022-05-30 NKI-AI/direct#205 @georgeyiasemis
Thanks for he comments @osorensen! Fixed
