### 2022-05-30 #4278 @georgeyiasemis
> @georgeyiasemis, the reviewers have now completed their checklists. I will now read through the paper, and get back to you if I have any suggested changes.
>
> In the meantime, could you:
>
> * [ ]  Make a tagged release of your software, and list the version tag of the archived version here (v1.0.1).
> * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository) (https://zenodo.org/record/6594702#.YpTPCmjMJL8)
> * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
> * [ ]  Please list the DOI of the archived version here. (DOI: **10.5281/zenodo.6594702**)
