+++ paper.bib
@@ -30,22 +30,25 @@ URL = {https://doi.org/10.1117/12.2609876}
 }
 
 @misc{beauferris2020multichannel,
-      author = {Youssef Beauferris and Jonas Teuwen and Dimitrios Karkalousos and Nikita Moriakov and Mattha Caan and George Yiasemis and Lívia Rodrigues and Alexandre Lopes and Hélio Pedrini and Letícia Rittner and Maik Dannecker and Viktor Studenyak and Fabian Gröger and Devendra Vyas and Shahrooz Faghih-Roohi and Amrit Kumar Jethi and Jaya Chandra Raju and Mohanasankar Sivaprakasam and Mike Lasby and Nikita Nogovitsyn and Wallace Loos and Richard Frayne and Roberto Souza},
-      title = {Multi-Coil MRI Reconstruction Challenge -- Assessing Brain MRI Reconstruction Models and their Generalizability to Varying Coil Configurations},
-      year={2020},
-      eprint={2011.07952},
-      archivePrefix={arXiv},
-      primaryClass={eess.IV}
+  doi = {10.48550/ARXIV.2011.07952},
+  url = {https://arxiv.org/abs/2011.07952},
+  author = {Beauferris, Youssef and Teuwen, Jonas and Karkalousos, Dimitrios and Moriakov, Nikita and Caan, Mattha and Yiasemis, George and Rodrigues, Lívia and Lopes, Alexandre and Pedrini, Hélio and Rittner, Letícia and Dannecker, Maik and Studenyak, Viktor and Gröger, Fabian and Vyas, Devendra and Faghih-Roohi, Shahrooz and Jethi, Amrit Kumar and Raju, Jaya Chandra and Sivaprakasam, Mohanasankar and Lasby, Mike and Nogovitsyn, Nikita and Loos, Wallace and Frayne, Richard and Souza, Roberto},
+  title = {Multi-Coil MRI Reconstruction Challenge -- Assessing Brain MRI Reconstruction Models and their Generalizability to Varying Coil Configurations}, 
+  publisher = {arXiv},
+  year = {2021},
+  copyright = {arXiv.org perpetual, non-exclusive license}
 }
 
 @misc{putzky2019irim,
-      title={i-RIM applied to the fastMRI challenge},
-      author={Patrick Putzky and Dimitrios Karkalousos and Jonas Teuwen and Nikita Miriakov and Bart Bakker and Matthan Caan and Max Welling},
-      year={2019},
-      eprint={1910.08952},
-      archivePrefix={arXiv},
-      primaryClass={eess.IV}
+  doi = {10.48550/ARXIV.1910.08952},
+  url = {https://arxiv.org/abs/1910.08952},
+  author = {Putzky, Patrick and Karkalousos, Dimitrios and Teuwen, Jonas and Miriakov, Nikita and Bakker, Bart and Caan, Matthan and Welling, Max},
+  title = {i-RIM applied to the fastMRI challenge},  
+  publisher = {arXiv}, 
+  year = {2019},
+  copyright = {arXiv.org perpetual, non-exclusive license}
 }
+
 
 @inbook{varnetfastmri,
 author = {Sriram, Anuroop and Zbontar, Jure and Murrell, Tullie and Defazio, Aaron and Zitnick, C. and Yakubova, Nafissa and Knoll, Florian and Johnson, Patricia},
@@ -121,6 +124,7 @@ year = {2018}
 
 @misc{zbontar2019fastmri,
       title={fastMRI: An Open Dataset and Benchmarks for Accelerated MRI},
+      doi = {10.48550/ARXIV.1811.08839},
[…]
