+++ paper.md
@@ -37,7 +37,7 @@ bibliography: paper.bib
 
 # Summary
 
-DIRECT is a Python, end-to-end pipeline for solving Inverse Problems emerging in Image Processing. It is built with PyTorch [@NEURIPS2019_9015] and stores state-of-the-art Deep Learning imaging inverse problem solvers for solving inverse problems such as denoising, dealiasing, and reconstruction. By defining a base forward linear or non-linear operator, DIRECT can be used for training models for recovering images such as MRIs from partially observed or noisy input data. Additionally, it provides the user with the functionality to load saved weights of pre-trained models to be used for inference. Furthermore, it offers functions for preparing and pre-processing data such as `.h5` files into PyTorch Datasets compatible with the software's training pipeline, but also allows for flexibility to work with any kind of PyTorch Dataset. Additionally, in order for the user to view the process of their experiments, it allows for continuous visualisation of training and validation metrics as well as image predictions utilising Tensorboard (examples are illustrated in Figures 1 and 2). 
+DIRECT is a Python, end-to-end pipeline for solving inverse problems emerging in image processing. It is built with PyTorch [@NEURIPS2019_9015] and stores state-of-the-art deep learning imaging inverse problem solvers for solving inverse problems such as denoising, dealiasing, and reconstruction. By defining a base forward linear or non-linear operator, DIRECT can be used for training models for recovering images such as MRIs from partially observed or noisy input data. Additionally, it provides the user with the functionality to load saved weights of pre-trained models to be used for inference. Furthermore, it offers functions for preparing and pre-processing data such as `.h5` files into PyTorch datasets compatible with the software's training pipeline, but also allows for flexibility to work with any kind of PyTorch dataset. Additionally, in order for the user to view the process of their experiments, it allows for continuous visualisation of training and validation metrics as well as image predictions utilising Tensorboard (examples are illustrated in Figures 1 and 2). 
 
 | ![image](https://user-images.githubusercontent.com/71031687/138093195-67004ec7-6bfd-448b-ba53-4cdd291a471b.png) |
 |:--:|
@@ -49,7 +49,7 @@ DIRECT is a Python, end-to-end pipeline for solving Inverse Problems emerging in
  
 # Statement of need
 
-A plethora of image processing problems arising in biology, chemistry and medicine can be defined as inverse problems. Inverse problems aim in recovering a signal $\vec{x} \, \in \, \mathcal{X}$ (e.g. an image) that cannot be directly observed from a set of measurements $\vec{y} \, \in \, \mathcal{Y}$ and is subject to a given corruption process known as the forward model:
+A plethora of image processing problems arising in biology, chemistry, and medicine can be defined as inverse problems. Inverse problems aim in recovering a signal $\vec{x} \, \in \, \mathcal{X}$ (e.g. an image) that cannot be directly observed from a set of measurements $\vec{y} \, \in \, \mathcal{Y}$ and is subject to a given corruption process known as the forward model:
     
 \begin{equation}
     \vec{y} \, = \, \mathcal{A}(\vec{x}) \,+\,\vec{n},
@@ -65,7 +65,7 @@ where $\mathcal{A}$ denotes the forward operator and $\vec{n}$ is some measureme
 
 ## Accelerated Parallel MRI Reconstruction
 
[…]
