### 2022-05-30 #4278 @georgeyiasemis
> 👋 @georgeyiasemis - I'm the AEiC on duty this week, and I have a few small changes for the paper before we publish it. Please merge [NKI-AI/direct#206](https://github.com/NKI-AI/direct/pull/206) or let me know what you disagree with.

Hi @danielskatz. Thanks for that. Merged it !
